% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data retrieved from AWQMS}
\usage{
clean_data(
  data,
  dql = c("A", "B", "E"),
  result_status = c("Final", "Provisional")
)
}
\arguments{
\item{data}{The result of an Oregon DEQ AWQMS query.}

\item{dql}{Vector of acceptable data quality levels. Can be any of 'A', 'B', 'C', 'D', E', 'F', or NA}

\item{result_status}{Vector of acceptable results tats. Can be any of 'Final', 'Preliminary', 'Rejected', 'Validated', or 'Provisional'.}
}
\value{
A slimmer data frame checked for quality.
}
\description{
Removes unnecessary variables, checks data for quality,
converts ug/L to mg/L and degrees Fahrenheit to degrees Celsius, adds
"sample_datetime" and "sample_id" columns. Updates 'Temp_Criteria' column to 13 during spawning period.
Sorts data by 'MLocID', 'Char_Name', and 'sample_datetime' columns
See DEQ09-LAB-006-QAG; Version 3.0 and https://www.oregon.gov/deq/wq/Documents/wqmAWQMSguide.pdf for explanation of DQLs and result status.
}
