% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impaired_waters.R
\name{impaired_waters}
\alias{impaired_waters}
\title{impaired_waters}
\usage{
impaired_waters(database, parameter = NULL, subbasins = NULL)
}
\arguments{
\item{database}{Source database}

\item{parameter}{string, or vector of strings, for parameter to be used}

\item{subbasins}{string, or vector of strings, of HUC 4 subbasin names}
}
\value{
Dataframe of IR category 4 and category 5
}
\description{
This function will return a dataframe of waterbodies listed as either category 4 or 5 for
a given set of subbasins (huc4's) and parameters
}
