% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strds_geo_ref_lookup.R
\name{standards_lookup}
\alias{standards_lookup}
\title{standards_lookup}
\usage{
standards_lookup(
  HUC8,
  exclude_field = NULL,
  exclude_params = NULL,
  layer = "Oregon_Standards",
 
    gdb_file = "//deqhq1/GISLIBRARY/Base_Data/DEQ_Data/Water_Quality/WQ_Standards/GeoRef_Standards.gdb"
)
}
\arguments{
\item{HUC8}{Vector of HUC8s to filter on.}

\item{layer}{Which layer of geodatabase to use, Defaults to "Oregon_Standards"}

\item{gdb_file}{Location of the gdf folder to use}
}
\value{
Dataframe of data from the standards layer
}
\description{
Exports data from the georeferenced standards database. Geodatabase is
fairly large, so input requires a vector if HUC8 codes to use to filter
data. Function will return data table of data from those HUC8s
}
\examples{
 \dontrun{
  standards_lookup(HUC8 = c("17050116", "17050117", "17050118", "17050119", "17050115"))
  }
}
